+++ electricitylci/egrid_filter.py
@@ -1,5 +1,5 @@
 ###Creates the data for electricity generation processes by fuel type and eGRID subregion
-###Uses global variables set in the globals file that define filters
+###Uses global variables set in the globals file that define fifile:///C:/Users/TGhosh/Dropbox/Electricity_LCI/2. Electric_Refactor after July/ElectricityLCI/electricitylci/%23%23_____code%23for%23developer.pylters
 import warnings
 warnings.filterwarnings("ignore")
 
@@ -15,9 +15,9 @@ from electricitylci.egrid_energy import list_egrid_facilities_in_efficiency_rang
 from electricitylci.egrid_emissions_and_waste_by_facility import emissions_and_wastes_by_facility
 from electricitylci.egrid_emissions_and_waste_by_facility import years_in_emissions_and_wastes_by_facility
 from electricitylci.egrid_FRS_matches import list_FRS_ids_filtered_for_NAICS
-from electricitylci.egrid_database_generator import create_process_dict
-from electricitylci.egrid_template_generator import *
-from electricitylci.globals import efficiency_of_distribution_grid
+from electricitylci.egrid_generation_database_builder import create_process_dict
+from electricitylci.egrid_template_builder import *
+
 
 
 
@@ -58,19 +58,12 @@ len(egrid_facilities_to_include)
 #Start with all emissions and wastes; these are in this file
 emissions_and_waste_by_facility_for_selected_egrid_facilities = emissions_and_wastes_by_facility[emissions_and_wastes_by_facility['eGRID_ID'].isin(egrid_facilities_to_include)]
 #for egrid 2016,TRI 2016,NEI 2016,RCRAInfo 2015: 118415
-
+#emissions_and_waste_by_facility_for_selected_egrid_facilities.to_excel('Mainfile.xlsx')
 generation_process_dict,generation_mix_dict = create_process_dict(emissions_and_waste_by_facility_for_selected_egrid_facilities)
-
-#The distribution mix dictionary does not require any new information and it will be fine to use the GEN mix dictionary to write these templates.
-#Only extra infomration required is the efficiency
-distribution_dict = generation_mix_dict
-
 
 gen_process_template_generator(generation_process_dict)
 
 gen_mix_template_generator(generation_mix_dict)
-
-distribution_template_generator(distribution_dict,efficiency_of_distribution_grid)
 
 
 '''
