+++ electricitylci/eia923_generation.py
@@ -8,7 +8,7 @@ import requests
 
 def eia_download_extract(odd_year):
         
-        data_dir = os.path.dirname(os.path.realpath(__file__))+"\\data\\"
+        data_dir = os.path.dirname(os.path.realpath(__file__))+"/data/"
         os.chdir(data_dir)
